refactor: organize data layer and domain layer structure. (#16)
