+++ app/src/main/java/amin/codelabs/qdo/data/task/dao/TaskDao.kt
@@ -1,5 +1,6 @@
-package amin.codelabs.qdo.data.task
+package amin.codelabs.qdo.data.task.dao
 
+import amin.codelabs.qdo.data.task.entity.TaskEntity
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
