+++ app/src/main/java/amin/codelabs/qdo/domain/task/usecase/DeleteTaskUseCase.kt
@@ -1,5 +1,6 @@
-package amin.codelabs.qdo.domain.task
+package amin.codelabs.qdo.domain.task.usecase
 
+import amin.codelabs.qdo.domain.task.repository.TaskRepository
 import javax.inject.Inject
 
 /**
